Moved command Login class to user.py
